Created CTime class for framerate timing
Created the CTime class for framerate timing based on the old framrate
class.
In truth, this new CTime class is nearly identical to the old framerate
class, however, it has been written in a neater fashion, and with
fuller comments (and it gave me a refresher on how the framerate class
worked).
Changed the dependent classes such as EngineInstance to be compatible
with the new CTime class as opposed to the old framerate class.
Cleaned up some Java warnings like unused imports and generated
serialisation UID's.