CTime Bug Fixes
Fixed major bugs in CTime that caused the very first framerate reported
to be inaccurate.
Fixed a separate bug to prevent the framerate from ever equalling 0;
although this fix now means that the reported framerate is always 1
more than the true framerate. This inaccuracy, however, is not a big
issue.