Changed RealFramerate to Float instead of Int
When the framerate is lower this effects all of the entities in the
engine by causing their overall movement speed to be slower. In an
attempt to remedey this bug the real framerate was changed from an
integer to a float for increased accuracy. (since a small difference
can be significant when amplified due to multiple frames per second.)
Unfortunately, the problem still occurs. It may be unavoidable,
although shouldn't really matter, as EVERY entity in the game is slowed
by the same amount. This bug should be fixed before the "release" of
the engine though.